car/requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
django-modeltranslation==0.19.13
django-multiselectfield==0.1.13
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
phonenumbers==9.0.3
phonenumberslite==9.0.3
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
